requirements.txt
# OpenAI API
openai>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilities
numpy>=1.24.0
